Refactor style components into style class
