+++ nachomemes/__init__.py
@@ -1,6 +1,6 @@
-
 from .guild_config import GuildConfig
-from .store import Color, Font, Justify, TextBox, MemeTemplate, Store, da_config, TemplateError, guild_id
+from .style import Color, Font, Justify, TextBox
+from .store import MemeTemplate, Store, da_config, TemplateError, guild_id
 from .dynamo_store import DynamoTemplateStore
 from .local_store import LocalTemplateStore
 from .render import render_template
