+++ nachomemes/store.py
@@ -1,5 +1,3 @@
-# pylint: skip-file
-
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from enum import Enum, auto
@@ -15,6 +13,7 @@ from dacite import Config
 from PIL import Image, ImageFont
 
 from .guild_config import GuildConfig
+from .style import Color, Justify, Font, TextBox
 
 # Monkeypatch Request to show the url in repr
 Request.__repr__ = lambda self: f"Request(<{self.full_url}>)"
@@ -46,62 +45,6 @@ def _fetch_image(url: Request) -> IO:
 
 class TemplateError(Exception):
     pass
-
-
-class Color(Enum):
-    BLACK = (0, 0, 0)
-    WHITE = (255, 255, 255)
-
-
-class Justify(Enum):
-    """Horizontal text position; lambda function calculates left offset from
-    enclosing box"""
-
-    LEFT = (lambda w1, w2: 0,)
-    CENTER = (lambda w1, w2: (w1 - w2) // 2,)
-    RIGHT = (lambda w1, w2: w1 - w2,)
-
-    def __call__(self, *args, **kwargs):
-        return self.value[0](*args, **kwargs)
-
-
-class Font(Enum):
-    IMPACT = Path("fonts/impact.ttf")
-    XKCD = Path("fonts/xkcd-script.ttf")
-    COMIC_SANS = Path("fonts/comic.ttf")
-
-    def load(self, font_size: int) -> ImageFont:
-        return ImageFont.truetype(str(self.value), font_size)
-
-
-@dataclass
-class TextBox:
-    "Definition for text that will be placed in a template."
-
-    # left, right, top and bottom are offsets from the top left corner as a
-    # percentage of the target image
-    left: float
-    right: float
-    top: float
-    bottom: float
-
-    face: Font
-
-    # in pixels
-    max_font_size: Optional[int]
-
-    color: Color = Color.BLACK
-    outline: Optional[Color] = None
-
-    # text alignment
-    justify: Justify = Justify.CENTER
-
-    # in degrees
-    rotation: Optional[int] = 0
-
-    # if this textbox is sized independently of the other boxes
-    ind_size: Optional[bool] = False
-
 
 @dataclass
 class MemeTemplate:
